+++ test/bigtest/network/factories/instance.js
@@ -13,6 +13,7 @@ export default Factory.extend({
   title: () => lorem.sentence(),
   contributors: () => [{ name: `${name.lastName()}, ${name.firstName()}` }],
   source: () => 'FOLIO',
+  matchKey: () => 'match_key',
   identifiers: () => [],
   publication: () => [],
   alternativeTitles: () => [],